Add primary TSEM implementation file.
The tsem.c file is the 'master' file in the TSEM implementation.
It is responsible for initializing the LSM and providing
implement of the security event hooks.

In addition to initializing the LSM, the set_ready() function
implements a secondary initialization that is used to a
statically scoped variable that indicates that security event
modeling can begin.  This is required secondary to the fact that
the cryptographic API's do not become ready until after the 'fs'
phase of system initialization is complete.

This file also handles the implementation of the tsem_mode kernel
command-line option.  If tsem_mode is set to a value of 1 the
TSEM driver does not model any events from the root domain.  This
is intended to allow development platforms to develop security
models without the overhead of full platform modeling.

This file also contains the implementation of the tsem_names
array that contains the ASCII text names that are assigned to
each security event handler.  This name is used as one of the
characteristics in the security state points that are generated.
This array is also used to provide symbolic names for the export
of security event descriptions, either through the tsemfs
filesystem or for export to external trust orchestrators.

Signed-off-by: Greg Wettstein <greg@enjellic.com>