Use the proper TMA authentication key.
A regression was noted with a kernel segmentation fault on the
first exported event to an external TMA.

The segmentation fault was being caused by attempting to use the
external modeling context after the failure to setup the process
authentication key.

The failure to setup the authentication key was caused by the
fact that the tsem_ns_event_key() function was being called with
an unused character buffer that had been allocated on the stack
in an earlier implementation.  The pointer to that buffer was
replaced with the pointer to the ASCII representation of the TMA
authentication key.

This then lead to an investigation of why the error return was
not being propagated to userspace.  This lead to what appears to
be some type of odd behavior from the hex2bin() function.

That function should return a negative error code if an incorrect
character is encountered in the input buffer to the function.
Instead the function is return an integer value of one.  This
value is not interpreted as an error return leading to the
control call succeeding when the initialization did not properly
complete.

This problem was fixed by setting the return value to -EINVAL in
the case of any non-zero return from the hex2bin function.