Allow selection of user namespace reference.
Previously the initial user namespace was used as the reference
for the user identity values used in the internal model or
exported to userspace.  On initial review Paul Moore had
indicated that it should be emphasized what reference space was
used for the uid/gid values.

After some reflection it appeared the best way forward would be
to make this option selectable when a modeling domain was
created.  This commit provides the infrastructure necessary to
implement this function.

The control message to create a namespace was modified to accept,
as the second argument a keyword to describe what namespace
should be used for the uid/gid reference.  The keyword 'init'
causes the modeling domain to use the initial user namespace.
The keyword 'current' causes the modeling domain to use the
current user namespace of the process executing the security
event.

The syntax for the control messages are now as follows:

internal [init|current]

external [init|current] KEY

The fs.c:write_control() function was modified to use the
following helper function to parse the control message:

fs.c:config_namespace()

This is in anticipation of adding a keyword to select what hash
function is to be used for the domain.

The following enumeration type was introduced to report the two
states of the uid/gid mapping:

tsem_ns_config

The namespace.c:tsem_ns_create() function was updated to allow
this enumeration type to be passed to the namespace creation
infrastructure in order to select the uid/gid reference for the
domain.

The tsem_TMA_context structure that defines the context for a
trusted modeling agent was modified to contain the following
structure member:

use_current_ns

That is set to a boolean value of true in order to cause the code
in event.c that creates the event description structure to use
the current user namespace rather than the initial user
namespace.

The fs.c:show_event() and fs.c:show_file() functions were
modified to use an unsigned printf format specifier for the
uid/gid fields.  This was done in order to prevent the export of
a -1 value as an overflow id for a uid/gid value that does not
have a mapping in a user namespace.

Since the signature of the namespace:tsem_create_ns() function
was modified to pause the pointer to the TMA control key to a
const char *.  This signature change was propagated downward to
the actual use of the character string.

A check for a valid length of the TMA control key was implemented
in the fs.c:config_namespace() function.