Implement framework for configurable digest functions.
One of the observations from the initial code review was that
TSEM was using only SHA256 as the compression function.  The
recommendation was to have the ability to configure the hash
function that was being used.

No specific recommendations had been given with respect to which
hash functions should be made available or the length of the
digest.

After some experimentation it became apparent that it will be
possible to support any hash function that the kernel implements.
The will be implemented by passing the name that the kernel uses
to internally identify the hash function as an argument to the
namespace creation command.

In preparation for making the digest function be selectable the
code was updated to eliminate all dependencies that had been
created by the code assuming that SHA256 would be available.

The tsem_TMA_context structure was updated to include the
following two new elements:

char *digest

unsigned int digestsize

The digest variable contains a pointer to a character array that
holds the name of the digest function that will be used.  The
digestsize element will contain the length in bytes of the digest
value that the specified hash will produce.

The following two inline functions were also introduced to return
the name of the hash function and the digest size:

tsem_digest()

tsem_digestsize()

All references to the "sha256" string were replaced with a call
to the tsem_digest() function.  All references to the
SHA256_DIGEST_SIZE macro were changed to the tsem_digestsize()
function, with the exception of where the macro was used to
define a function local u8 array, in which case the
HASH_MAX_DIGESTSIZE macro was used.

The generic_cell value and the pseudonym digest arrays were
modified to be a total length of 64 bytes, the value of
HASH_MAX_DIGESTSIZE.  When these values are needed the code
copies a digestsize amount of the array into the compression
function.

The code was verified to produce the same point values for a
known execution trajectory.