Handle cacheing of multiple digest values.
Allowing multiple simultaneous implementations of hasing
functions for security modeling namespaces surfaced an issue with
the current file digest cacheing that stores a single digest
value in the tsem_inode structure.  The value cached was the last
digest value that was computed for the inode, which may not be
the proper digest value for a modeling domain using an alternate
hash function.

In order to address this the following global structure was
added:

tsem_inode_digest

A linked list of these structures, anchored in the tsem_inode
structure holds all of the various digest values that were
computed for an inode.  The name of the hash function that
generated the digest value is stored in the tsem_inode_digest
structure and is used to locate a relevant cached digest value
for a modeling domain.

Each tsem_inode_digest structure contains its own inode version
number so that computation of a digest value can be triggered if
the hash function is used in a modeling domain that has been
updated by a modeling domain using an alternate hash function.

In order to release memory consumed by the tsem_inode_digest
structures the following additional LSM hook was implemented by
TSEM:

inode_free_security

When an inode is unlinked, or its security structure is released,
the list of digests attached to the inode are walked and the
memory for the digest name and the structure itself are freed.

This change revealed a problem that will need to be subsequently
addressed with respect to the handling of a socket_accept event.

The digest value, previously stored in the tsem_inode structure,
was being used as a characteristic in the computation of the CELL
identity for the event.  This digest was not being explicitly
computed for a socket_accept value and was thus not
characteristic of the event, although it appeared to generate a
constant file.

A byte array was added to the tsem_socket_accept_args structure
to hold the digest value for what would currently be a non-IPV4
or non-IPV6 socket, the most common of which is an AF_UNIX
socket.  This byte array is explicitly set to zero and will be
properly populated in a subsequent update to this encounter.